cookbook/requirements.txt
attrs==22.2.0
click==8.1.3
coverage==7.0.1
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
Flask==2.2.2
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
PyYAML==6.0
tomli==2.0.1
Werkzeug==2.2.2
